ULTIMA8: Fix lots more coverity issues

+++ engines/ultima/ultima8/world/actors/actor.cpp
@@ -995,7 +995,8 @@ void Actor::clearInCombat() {
 	if ((_actorFlags & ACT_INCOMBAT) == 0) return;
 
 	CombatProcess *cp = getCombatProcess();
-	cp->terminate();
+	if (cp)
+		cp->terminate();
 
 	clearActorFlag(ACT_INCOMBAT);
 }
